requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==2.2.20
Flask==1.1.2
IbPy2==0.8.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.0
pandas==1.1.5
PyJWT==1.7.1
pyodbc==4.0.30
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.1
twilio==6.51.0
urllib3==1.26.2
websockets==8.1
Werkzeug==1.0.1
